% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_csv.R
\name{bind_csv}
\alias{bind_csv}
\title{Import PRIDICT2 predictions (.csv)}
\usage{
bind_csv(x)
}
\arguments{
\item{input_directory}{A directory.}
}
\description{
binds all PRIDICT output files into one data table from .csvs
following classes:
}
