% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameColValidations.R
\name{frameColVal_greaterThan0}
\alias{frameColVal_greaterThan0}
\title{frameColVal_greaterThan0}
\usage{
frameColVal_greaterThan0(frame, cols)
}
\description{
frameColVal_greaterThan0
WORKS_WITH_CTYPE: N
}

