% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameColValidations.R
\name{frameColVal_hasGenderLevelsEs}
\alias{frameColVal_hasGenderLevelsEs}
\title{frameColVal_hasGenderLevelsEs}
\usage{
frameColVal_hasGenderLevelsEs(frame, cols)
}
\description{
frameColVal_hasGenderLevelsEs
}

