% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{frameR_status}
\alias{frameR_status}
\title{Check the status of the frameR Python environment}
\usage{
frameR_status()
}
\value{
Called for its side effects. Prints status information.
}
\description{
Reports whether the frameR Python environment exists and
which Python packages are installed within it.
}
\examples{
\dontrun{
frameR_status()
}

}
