% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_PRIDICT.R
\name{import_PRIDICT}
\alias{import_PRIDICT}
\title{Import PRIDICT2 predictions (.csv)}
\usage{
import_PRIDICT(input_directory)
}
\arguments{
\item{input_directory}{A directory.}
}
\description{
binds all PRIDICT output files into one data table from .csvs
following classes:
}
