% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-Desktop.R
\name{print.frameR_results}
\alias{print.frameR_results}
\title{Print method for frameR_results}
\usage{
\method{print}{frameR_results}(x, ...)
}
\arguments{
\item{x}{A \code{frameR_results} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for frameR_results
}
