% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDf.R
\name{sampleData}
\alias{sampleData}
\title{Creates a new Dataframe from a dataframe}
\usage{
sampleData(ftype, asFrame = FALSE)
}
\arguments{
\item{d}{might be a json string, data frame or list of data frames.}
}
\value{
dp
}
\description{
Creates a new datapackage from json, data frame, list of data frames, list of data framees (see Dataframe reference class)
}
\examples{
\dontrun{
frame <- newDataframeFromDataframe(mtcars)
}
}

